Add Report to Patient/requirements.txt
python-docx
docx2pdf
pyrebase4
